docs: 주문 엔티티 설정 주문 엔티티 설정
